+++ pack/bundle/opt/fzf/install.ps1
@@ -1,4 +1,4 @@
-$version="0.73.0"
+$version="0.74.0"
 
 $fzf_base=Split-Path -Parent $MyInvocation.MyCommand.Definition